requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
sentence-transformers==2.3.1
PyPDF2==3.0.1
python-docx==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
faiss-cpu==1.7.4
groq==0.4.0
numpy==1.24.4